updating redocly config and plugin format
